requirements.txt
Flask==0.12
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Markdown==2.6.7
bs4==0.0.1
requests==2.12.4
